% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_calcs.R
\name{calc_hom}
\alias{calc_hom}
\title{Calculate The Homogeneity of Cell Gene Expression in scRNA-seq Data.}
\usage{
calc_hom(counts = NULL, normalise = FALSE, ...)
}
\arguments{
\item{counts}{The counts matrix. Rows must be cells and columns must be genes.}

\item{normalise}{Whether the homogeneity should be returned as a proportion of
the maximum possible homogeneity.}

\item{...}{Arguments passed to \code{entropy::entropy()}. It is recommended these are
not included when using \code{normalise = TRUE}.}
}
\value{
A vector of homogeneity scores.
}
\description{
Use entropy methods from Michael J. Casey to quantify homogeneity using
information theory.
}
\examples{

mat <- matrix(1:12, ncol = 3)
calc_hom(mat, normalise = TRUE)

}
